% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{get_all_categories}
\alias{get_all_categories}
\title{Get all available browser/crawler categories from useragentstring.com}
\usage{
get_all_categories()
}
\value{
A tibble with columns: name (browser/crawler name), category (CRAWLERS, BROWSERS, etc.)
}
\description{
Get all available browser/crawler categories from useragentstring.com
}
\examples{
\dontrun{
categories <- get_all_categories()
head(categories)
}
}
